pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-checker"
version = "0.1.0"
description = "Tägliche Portfolio-Zusammenfassung (Broker-Connectoren)"
requires-python = ">=3.11"
dependencies = [
    "pyetrade>=2.1.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
portfolio-checker = "portfolio_checker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
